requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
pandas
python-dotenv
requests